refactored everything to double to fix weird graphical problem

+++ src/dda_algo/dda_move.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   dda_move.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: psanger <psanger@student.42.fr>            +#+  +:+       +#+        */
+/*   By: nburchha <nburchha@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/26 17:12:59 by psanger           #+#    #+#             */
-/*   Updated: 2024/06/26 17:23:11 by psanger          ###   ########.fr       */
+/*   Updated: 2024/07/01 17:27:30 by nburchha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
-int	move_x(t_dda *dda, char **map, float hyplen_x)
+int	move_x(t_dda *dda, char **map, double hyplen_x)
 {
 	dda->end_x = dda->end_x + dda->delta_x;
 	dda->end_y = dda->end_y + (hyplen_x * fabs(sin(dda->angle))
@@ -22,7 +22,7 @@ int	move_x(t_dda *dda, char **map, float hyplen_x)
 	return (0);
 }
 
-int	move_y(t_dda *dda, char **map, float hyplen_y)
+int	move_y(t_dda *dda, char **map, double hyplen_y)
 {
 	dda->end_y = dda->end_y + dda->delta_y;
 	dda->end_x = dda->end_x + (hyplen_y * fabs(cos(dda->angle))
@@ -34,8 +34,8 @@ int	move_y(t_dda *dda, char **map, float hyplen_y)
 
 int	get_new_pos(t_dda *dda, char **map)
 {
-	float	hyplen_x;
-	float	hyplen_y;
+	double	hyplen_x;
+	double	hyplen_y;
 
 	hyplen_x = 0;
 	hyplen_y = 0;
